requirements.txt
streamlit==1.39.0
requests>=2.31.0
pydantic>=2.7.0
python-dotenv>=1.0.1
langgraph>=0.2.39
langchain-core>=0.3.15
pydeck>=0.9.1
ics>=0.7.2
pandas>=2.2.2
geopy>=2.4.1
